Probability System Essembled. Only missing the color system.
The initial plans for a color based probability system using shaders fell through due to the shaders not performing as expected with pixel art. So we switch to a new simpler system using a main sprite and a separate 'aura sprite' for the color.

The probability system is also revamped. The ranges of probabilities that were represented by the colors are replaced by the simpler more readable whole values representing the probabilities.